requirements.txt
python-telegram-bot==20.7
opencv-python-headless==4.9.0.80
face-recognition==1.3.0
numpy<2.0
